ui-charts: move some common hooks and types in /common
