+++ front/ui/ui-charts/src/common/context.ts
@@ -1,0 +1,12 @@
+import { createContext } from 'react';
+
+import type { CanvasContextType } from './types';
+
+/**
+ * This context only contains necessary helpers to (un)register drawing functions.
+ * It is basically never updated.
+ */
+export const CanvasContext = createContext<CanvasContextType>(
+  // That value should never be used, since the context should always be accessed within a provider
+  undefined as unknown as CanvasContextType
+);
