+++ front/ui/ui-charts/src/common/index.ts
@@ -1,0 +1,5 @@
+export { CanvasContext } from './context';
+
+export { usePicking, useDraw } from './useCanvas';
+
+export type { HoveredItem, DrawingFunction, PickingDrawingFunction, PickingElement } from './types';
