+++ front/ui/ui-charts/src/index.ts
@@ -1,5 +1,7 @@
 import '@osrd-project/ui-core/dist/theme.css';
 
+export * from './common';
+
 export * from './spaceTimeChart';
 
 export * from './manchette';
