+++ front/ui/ui-charts/src/spaceTimeChart/components/OccupancyBlockLayer.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 
-import { useDraw } from '../hooks/useCanvas';
-import { type DrawingFunction } from '../lib/types';
+import type { DrawingFunction } from '../../common/types';
+import { useDraw } from '../../common/useCanvas';
 
 export type OccupancyBlock = {
   timeStart: number;
