+++ front/ui/ui-charts/src/spaceTimeChart/lib/context.ts
@@ -1,10 +1,6 @@
 import { createContext } from 'react';
 
-import {
-  type CanvasContextType,
-  type MouseContextType,
-  type SpaceTimeChartContextType,
-} from './types';
+import { type MouseContextType, type SpaceTimeChartContextType } from './types';
 
 // There are three different contexts because they have very different lifecycles.
 
@@ -15,15 +11,6 @@ import {
 export const SpaceTimeChartContext = createContext<SpaceTimeChartContextType>(
   // That value should never be used, since the context should always be accessed within a provider
   undefined as unknown as SpaceTimeChartContextType
-);
-
-/**
- * This context only contains necessary helpers to (un)register drawing functions.
- * It is basically never updated.
- */
-export const CanvasContext = createContext<CanvasContextType>(
-  // That value should never be used, since the context should always be accessed within a provider
-  undefined as unknown as CanvasContextType
 );
 
 /**
